Fixed zoom greater than 100%
In commit 84219bd84cfd9935f2be7c9055d28a4446ea2260, an optimization was
introduced for zoom greater than 100%.  That commit refactored a block of
code out into a separate function.  The function param "tileImg" was
partially renamed to "imageElement", but not entirely.

This commit changes all uses of "imageElement" back to "tileImg".